another bugfix
git-svn-id: http://svn.resourcespace.org/svn/resourcespace@1220 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/image_processing.php
@@ -506,8 +506,10 @@ function create_previews_using_im($ref,$thumbonly=false,$extension="jpg",$previe
 			$file=get_resource_path($ref,true,"tmp",false,"jpg");	
 			}
 
-		$hpr_path=get_resource_path($ref,true,"hpr",false);	
-		if (file_exists($hpr_path)) {unlink($hpr_path);}	
+		if ($previewonly || $previewbased){
+			$hpr_path=get_resource_path($ref,true,"hpr",false);	
+			if (file_exists($hpr_path)) {unlink($hpr_path);}	
+		}
 		$lpr_path=get_resource_path($ref,true,"lpr",false);	
 		if (file_exists($lpr_path)) {unlink($lpr_path);}	
 		$scr_path=get_resource_path($ref,true,"scr",false);	
@@ -554,7 +556,7 @@ function create_previews_using_im($ref,$thumbonly=false,$extension="jpg",$previe
 			$tw=$ps[$n]["width"];$th=$ps[$n]["height"];
 			$id=$ps[$n]["id"];
 
-			if ((!$highestsize && !eregi("jp[e]?g", $extension)) || ($sw>$tw) || ($sh>$th) || ($id == "pre") || ($id=="thm") || ($id=="col") || $previewbased || ($previewonly && $id=="scr"))
+			if ((!$highestsize && !eregi("jp[e]?g", $extension)) || ($sw>$tw) || ($sh>$th) || ($id == "pre") || ($id=="thm") || ($id=="col") || ($previewonly && $id=="scr"))
 				{
 				if (($sw<$tw) && ($sh<$th))
 					{
